requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
fonttools==4.29.1
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
fonttools==4.29.1
idna==3.3
ipython==8.1.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
stack-data==0.2.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
